refactor: move series information to a new model that is shared
